Rename application files; update routes.
