--- src/PresentationalComponents/SampleComponent/sample-component.scss
@@ -1,7 +1,0 @@
-@import "~@redhat-cloud-services/frontend-components-utilities/files/Utilities/_all";
-
-.sample-component {
-  font-size: $ins-fontSize;
-  margin: $ins-margin;
-  display: block;
-}
